pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "trading-agent-polymarket"
version = "0.1.0"
description = "Multi-Agent LLM Framework for Polymarket Prediction Market Analysis"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain-core>=0.3.81",
    "langchain-anthropic>=0.3.15",
    "langchain-google-genai>=2.1.5",
    "langchain-openai>=0.3.23",
    "langgraph>=0.4.8",
    "python-dotenv>=1.0.0",
    "rank-bm25>=0.2.2",
    "requests>=2.32.4",
]

[tool.setuptools.packages.find]
include = ["tradingagents*"]
